==> zcta_map/1_GeoJSON_Layer.ipynb <==
# Step 1: Initial data work and extraction of geodata geometries
###### Geographic Data Cleaning and Assembly for Google Maps using Jupyter Gmaps;
------
Before starting the data cleaning process, we obtained a list of zip codes within the DC Metro area through a web utility attempting to select as many as possible zip codes that exist within the bounds of the D.C. Beltway.

Utilizing data published on census.gov, we start here with some simple organization and slimming down of the bulk of the dataset. Some of the datasets the team had found included data for all US zipcodes (the GeoJSON data comes to mind specifically).

Ultimately, before the conclusion of the project, much of this work would be implemented in more fully developed code in other notebooks. Given more time, this particular notebook would be revised if not outright retired. However, some specific output documents are still in use by other project components.


import pandas as pd
import json
import csv
from keys import gapikey
from matplotlib.cm import winter
from matplotlib.colors import to_hex
from area import area
import ast

with open('zcta/zcta.json') as a:
    geoj = json.load(a)
with open('resources/zip_targets.csv', newline='\n', encoding='utf-8-sig') as b:
    ziptgt = pd.read_csv(b)
with open('resources/dmv_pop.csv', newline='\n', encoding='utf-8-sig') as c:
    dc_pop = pd.read_csv(c)
# with open('resources/places_geo.csv', newline='\n', encoding='utf-8') as d:
#     places_geo = pd.read_csv(d)
with open('resources/median_income.csv', newline='\n', encoding='utf-8-sig') as f:
    dmv_income = pd.read_csv(f)

dmv_income.head()

dmv_incomedf = dmv_income[['GEO.id2', 'HC03_EST_VC02']].replace('-',0)
dmv_incomedf = dmv_incomedf.replace('250,000+',0)
dmv_incomedf = dmv_incomedf.rename(columns={'GEO.id2': 'zip', 'HC03_EST_VC02': 'median_household_income'}).drop(0).astype('int64')

dc_maindf = ziptgt.merge(dmv_incomedf, on='zip', how='outer').set_index('zip').fillna(value=0)

cu_id = dc_maindf[ dc_maindf['city'] == 0].index
dc_maindf = dc_maindf.drop(cu_id)
cu_id = dc_maindf[ dc_maindf['median_household_income'] == 0].index 
dc_maindf = dc_maindf.drop(cu_id)

dc_maindf.head()

# extract just population and zip from dc_pop.csv
dc_popdf = dc_pop[['GEO.id2','HC01_EST_VC01']]
# rename columns
dc_popdf = dc_popdf.rename(columns={'GEO.id2': 'zip', 'HC01_EST_VC01': 'population'}).drop(0).astype('int64')
# outer merge with target zipcode list, set index to zip and fill NaN
dc_maindf = dc_maindf.merge(dc_popdf, on='zip', how='outer').set_index('zip').fillna(value=0)

cu_id = dc_maindf[ dc_maindf['city'] == 0].index
dc_maindf = dc_maindf.drop(cu_id)
cu_id = dc_maindf[ dc_maindf['population'] == 0].index 
dc_maindf = dc_maindf.drop(cu_id)

zipdict = dict(dc_maindf['population'])

dc_maindf.head()

geoj_scrubbed = {}
areadict = {}
features = geoj['features']
dc_maindf['feature'] = 0

for feature in features:
    try:
        feature_geoid = int(feature['properties']['GEOID10'])
        if feature_geoid in dc_maindf.index.values.tolist():
            dc_maindf.loc[[feature_geoid],['feature']]= str(feature)
            zarea = area(feature['geometry'])/int(1E6)
            areadict.update({feature_geoid:zarea})
            
            
    except ValueError:
            print(f'Value Error')

areadf = pd.DataFrame.from_dict(areadict, orient='index', columns=['area'])
dc_maindf2 = dc_maindf.join(areadf, how='left')
dc_maindf2['density'] = dc_maindf2['population']/dc_maindf2['area']

dc_maindf2.head()

min_pop = min(dc_maindf2['population'])
max_pop = max(dc_maindf2['population'])
pop_range = max_pop - min_pop

dc_maindf2['nml_dens'] = ((dc_maindf2['population'] - min_pop)/pop_range)

min_inc = min(dc_maindf2['median_household_income'])
max_inc = max(dc_maindf2['median_household_income'])
inc_range = max_inc - min_inc

dc_maindf2['nml_inc'] = ((dc_maindf2['median_household_income'] - min_inc)/inc_range)

dc_maindf2.head()

feat_list = []
for feature in dc_maindf2['feature']:
    feat_list.append(ast.literal_eval(feature))

geoj_clean = {'type': 'FeatureCollection', 'features': feat_list}

with open('resources/features.json', 'w') as outfile:  
    json.dump(geoj_clean, outfile)
