==> zcta_map/2_Location_Query.ipynb <==
# Step 2: Location_Query
###### Let's find some healthcare locations... a lot of them. 
###### But for the love of all that is holy, do not run this again because running several thousand Google API calls can be EXPENSIVE. 
###### Thank goodness for the free trial credit is all I'm saying. Whew.
-----
Now we're going to cross reference some of the data we exported to CSV from step 1.

Using the centerpoint of the zipcodes we're visualizing, we will perform a nested loop of queries against those coordinates increasing our query range on each pass. (1km, 2km, and 3km). Once we have the returns, we use the Google API ID as a unique identifier to eliminate duplicates from the compiled list.

Through testing, it was found that assembling a master list by running multiple queries, using the google API next-page capability and removing duplicates after completion resulted in the most thourough query results.

As the query structure involves making several hundred api calls: (3 resolutions * 81 zipcodes * 2 api calls = 486), it does take time to run, and ultimately is being saved into a CSV format for later reference wihtout the need for unnecessary additional calls to the Google Places API. The query, run on April 22nd, 2019, returned 3820 unique results with type "Doctor".

Annoyingly, Google did not return those 3800 or so zipcodes along with the rest in the 'nearby' search results. This required a followup loop which pulls down location details on each item in the list by the place_id. This is then used to attach a list to the dataframe, allowing the doctors to be sorted by zipcode for later analysis. This data was exported to CSV.

Once doctors were done, the query was rerun with the place type as hospital. And those results exported as well.

With the healthcare location data extracted to CSV, this notebook largely only still exists to show work. I do not advise running this script without adequate credit with Google APIs.

import requests
import pandas as pd
import gmaps
import json
import csv
import time
import os
from keys import gapikey
from area import area
import ast
import re

with open('resources/zip_data.csv', newline='\n') as b:
    ziptgt = pd.read_csv(b)
with open('resources/zip_coords.csv', newline='\n', encoding='utf-8-sig') as c:
    zip_coords = pd.read_csv(c)

ziptgt = ziptgt.set_index('zip')
zip_coords = zip_coords.set_index('ZIP')
zip_df = ziptgt.join(zip_coords, how='left')

zip_df.head()

# !!! DO NOT RUN THIS AGAIN (THIS IS A VERY EXPENSIVE QUERY USE THE CSVs) !!!
# places_df = pd.DataFrame(columns=['geometry','icon','id', 'name', 'opening_hours', 'photos', 'place_id', 'plus_code', 'rating', 'reference', 'scope', 'types', 'user_ratings_total', 'vicinity'])

# i = 0
# print("Queries have started. Might be a good time to go get some coffee, it's going to take a few minutes")
# radius_list = [1000, 2000, 3000]
# for radius in radius_list:
#     print(f'Starting {radius} meter scan.')
#     for pzip in list(zip_df.index.values):
#         try:
#             npt = 0
#             target_geo = f"{zip_df['LAT'][pzip]},{zip_df['LNG'][pzip]}"
#             target_type = 'hospital'
#             target_radius = radius

#             params1 = {
#                 "location": target_geo,
#                 "radius": target_radius,
#                 "type": target_type,
#                 "key": gapikey,
#             }

#             base_url = "https://maps.googleapis.com/maps/api/place/nearbysearch/json"

#             response = requests.get(base_url, params=params1)

#             data1 = response.json()

#             npt = data1['next_page_token']

#             params2 = {
#                 "key": gapikey,
#                 "pagetoken": npt
#             }

#             time.sleep(2)

#             response = requests.get(f'https://maps.googleapis.com/maps/api/place/nearbysearch/json?key={gapikey}&pagetoken={npt}')

#             data2 = response.json()

#             places_df = places_df.append(data1['results'], ignore_index=True)
#             places_df = places_df.append(data2['results'], ignore_index=True)
#             i += 1
#             os.system('clear')
#             print(f'places retrieved for {pzip} | {i} of {len(zip_df.index)} @ {radius} meters')
#             print(f'{len(places_df.index)} places retrieved')
#         except KeyError:
#           print(f'key error, next zip')

#     places_df = places_df.drop_duplicates(subset='id').reset_index(drop=True)
    
# print('Query complete. Thanks for waiting.')

places_geo = pd.DataFrame(columns=['lat','lng','name','vicinity','placeid'])

for p in list(places_df.index.values):
    lat = places_df['geometry'][p]['location']['lat']
    lng = places_df['geometry'][p]['location']['lng']
    vicinity = places_df['vicinity'][p]
    name = places_df['name'][p]
    place_id = places_df['place_id'][p]
    geo_tmp = pd.DataFrame([[lat,lng,name,vicinity,place_id]],columns=['lat','lng','name','vicinity','placeid'])
    places_geo = places_geo.append(geo_tmp).reset_index(drop=True)

places_geo.head()

places_geo.to_csv('resources/hospital_geo_multi.csv', encoding='utf-8')

with open('resources/places_geo_multi.csv', newline='\n') as b:
    geo_places = pd.read_csv(b)
    
places_geo.head()

# !!! DO NOT RUN THIS AGAIN EITHER (THIS IS AN EVEN MORE EXPENSIVE QUERY USE THE CSVs) !!!
# place_zip = []
# expt_zip = []
# i = 1
# for placeid in places_geo['placeid']:
#     params3 = {
#         "placeid":placeid,
#         "key": gapikey,
#     }
#     mtchd = False
#     base_url = "https://maps.googleapis.com/maps/api/place/details/json"
#     response = requests.get(base_url, params=params3)
#     place_details = response.json()
#     print(f"processing {i} of {len(places_geo['placeid'])}")
#     i += 1
#     for addcomp in place_details['result']['address_components']:
#         match = re.match(r'^2\d{4}$', addcomp['short_name'])
#         if match:
#             if mtchd == False:
#                 place_zip.append(match.string)
#                 mtchd = True
#                 mtch1 = match.string
#             else:
#                 expt_zip.append(match.string)
#                 print(f"possible double match {match.string} & {mtch1}")

places_geo2 = places_geo
places_geo2['zip'] = place_zip

places_geo2.to_csv('resources/hospital_geo.csv', encoding='utf-8')
places_geo2.head()
