==> zcta_map/3_nml_and_color_v1.ipynb <==
# 3 - Normalized Score and Color Calculator
-----

This notebook was used to pare down a large amount of tabular census data to extract and calculate scores on the 81 zipcodes targeted for this project. By collecting some data preprocessed in either steps 1 or 2, we're able to marry that to data compiled from the US Census. This notebook outputs a primary data table for group analysis, as well as calculates normalized scores for specific values, then converts those values into color values based on the Matplotlib RdYlGn (Red Yellow Green) colormap. These color values will be used to populate the fill color values for the choropleth maps that will be generated in step 4.

import pandas as pd
from matplotlib.cm import RdYlGn
from matplotlib.colors import to_hex
import ast
import json

# Open your CSVs as a dataframes
    
with open('../all_econ_data.csv', newline='\n', encoding='utf-8-sig') as a:
    all_econ_data = pd.read_csv(a)
    
with open('../zcta_map/resources/zip_data.csv', newline='\n', encoding='utf-8-sig') as b:
    zip_data = pd.read_csv(b)
    
with open('../hcprovider_data/doctor_geo.csv', newline='\n', encoding='utf-8-sig') as c:
    doctor_geo = pd.read_csv(c)
    
with open('../hcprovider_data/hospital_geo.csv', newline='\n', encoding='utf-8-sig') as d:
    hospital_geo = pd.read_csv(d)
    
with open('../insurance_data/insurance_data.csv', encoding='utf-8') as e:
    ins_data = pd.read_csv(e)

all_econ_data.head()

zip_data.head()

doctor_geo.head()

ins_data.head()

dr_grpby = doctor_geo[['placeid','zip']].groupby('zip')
dr_df = pd.DataFrame(dr_grpby.count())

hos_grpby = hospital_geo[['placeid','zip']].groupby('zip')
hos_df = pd.DataFrame(hos_grpby.count())

dr_df.head()

hos_df.head()

ins_data1 = ins_data.drop(0).rename(index=str, columns={"GEO.id2.1": "zip", "HC03_VC131": "Percent Insured"}).drop(['GEO.id2'], axis=1)

ins_data1.head()

all_econ_data = all_econ_data.rename(index=str, columns={'Zip Code': 'zip'})

# all_econ_data = all_econ_data.set_index('zip')

all_econ_data.head()

econ_select = all_econ_data[['zip', 'Total Population', 'Unemployment Rate', 'Per Capita Income']]

econ_select.head()

ins_data2 = ins_data1
ins_data2['Percent Insured'] = ins_data2['Percent Insured'].replace('-',0)
ins_data2 = ins_data2.astype({'zip': 'int64', 'Percent Insured':'float64'})

ins_data2.head()

econ_select1 = econ_select.merge(ins_data2, on='zip', how='inner')

econ_spc = econ_select1.set_index('zip')

ziplist = list(zip_data['zip'])

econ_select1 = econ_spc.loc[ziplist]

econ_select1 = econ_select1.reset_index()

econ_select2 = econ_select1.merge(hos_df, on='zip', how='left')
econ_select2 = econ_select2.rename(index=str, columns={'placeid': 'hospital facilities'})
econ_select2 = pd.DataFrame(econ_select2)

econ_select2.head()

econ_select3 = econ_select2.merge(dr_df, on='zip', how='left')
econ_select3 = econ_select3.rename(index=str, columns={'placeid': 'doctors'})
econ_select3 = pd.DataFrame(econ_select3)

econ_select3.head()

def calculate_color(pop):
    mpl_color = RdYlGn(pop)
    gmaps_color = to_hex(mpl_color, keep_alpha=False)
    return gmaps_color

min_emp = min(econ_select3['Unemployment Rate'])
max_emp = max(econ_select3['Unemployment Rate'])
emp_range = max_emp - min_emp

econ_select3['nml_emp'] = (1 - (econ_select3['Unemployment Rate'] - min_emp)/emp_range)

econ_select3.head()

min_inc = min(econ_select3['Per Capita Income'])
max_inc = max(econ_select3['Per Capita Income'])
inc_range = max_inc - min_inc

econ_select3['nml_inc'] = ((econ_select3['Per Capita Income'] - min_inc)/inc_range)

inc_range

econ_select3['nml_ins'] = (econ_select3['Percent Insured'] / 100)

econ_select3.head()

colors = []

for pop in econ_select3['nml_inc'].tolist():
    try:
        color = calculate_color(pop)
    except KeyError:
        color = (0, 0, 0, 0.3)
        print('key error')
    colors.append(color)

econ_select3['color_key_inc'] = colors

econ_select3.head()

colors = []

for pop in econ_select3['nml_emp'].tolist():
    try:
        color = calculate_color(pop)
    except KeyError:
        color = (0, 0, 0, 0.3)
        print('key error')
    colors.append(color)

econ_select3['color_key_emp'] = colors

max(econ_select3['Per Capita Income'])

colors = []

for pop in econ_select3['nml_ins'].tolist():
    try:
        color = calculate_color(pop)
    except KeyError:
        color = (0, 0, 0, 0.3)
        print('key error')
    colors.append(color)

econ_select3['color_key_ins'] = colors

econ_select3.head()

zip_data.head()

econ_main = zip_data[['zip','feature','area','density']].rename(index=str, columns={'area': 'area km'})

econ_main = econ_main.merge(econ_select3, on='zip', how='left').fillna(0)

econ_main.to_csv('../primary_data.csv', encoding='utf-8')

feat_list = []
for feature in econ_main['feature']:
    feat_list.append(ast.literal_eval(feature))

geoj_clean = {'type': 'FeatureCollection', 'features': feat_list}

with open('../zcta_map/resources/features.json', 'w') as outfile:  
    json.dump(geoj_clean, outfile)
