==> zcta_map/4_Map_Gen_v1.ipynb <==
# Step 4 - Jupyter Gmaps Generator with HTML/CSS Legend Import

This script is responsible for building HTML files that will rendering the choropleth maps of the D.C. Metro. This notebook is importing a number of data sets collected throughout notebook steps 1 through 3.

Being imported are: GeoJSON geography dicts for choropleth boundary rendering, hexadecimal color keys generated through color mapping, Geographic coordinates of several thousand healthcare providers spread throughout the city, HTML/Style variables for rendering map legends through additional widgets. Outputs are to separate HTML files to improve notebook speed. (due to the very large number of rendered geographic points on the map, it is strongly advised to render a single map at a time).



import requests
import pandas as pd
import numpy as np
import gmaps
import json
import csv
from keys import gapikey
import legendhtml as lgnd
from ipywidgets.embed import embed_minimal_html
import ipywidgets as widgets
import matplotlib.pyplot as plt
gmaps.configure(api_key=gapikey)
from matplotlib.cm import RdYlGn
from matplotlib.colors import to_hex

with open('resources/places_geo2.csv', newline='\n', encoding='utf-8') as a:
    doctor_geo = pd.read_csv(a)

with open('resources/hospital_geo.csv', newline='\n', encoding='utf-8') as b:
    hospital_geo = pd.read_csv(b)
        
with open('resources/features.json') as c:
    geoj = json.load(c)
    
with open('../primary_data.csv', encoding='utf-8') as d:
    primary_data = pd.read_csv(d)

zcta_inc_colors = list(primary_data['color_key_inc'])
zcta_inc_layer = gmaps.geojson_layer(
    geoj,
    fill_color=zcta_inc_colors,
    stroke_color=zcta_inc_colors,
    fill_opacity=0.8,
    stroke_weight=0.5)

zcta_ins_colors = list(primary_data['color_key_ins'])
zcta_ins_layer = gmaps.geojson_layer(
    geoj,
    fill_color=zcta_ins_colors,
    stroke_color=zcta_ins_colors,
    fill_opacity=0.8,
    stroke_weight=0.5)

zcta_emp_colors = list(primary_data['color_key_emp'])
zcta_emp_layer = gmaps.geojson_layer(
    geoj,
    fill_color=zcta_emp_colors,
    stroke_color=zcta_emp_colors,
    fill_opacity=0.8,
    stroke_weight=0.5)

doctor_layer = gmaps.symbol_layer(doctor_geo[['lat','lng']], fill_color='black', stroke_color='black', scale=1)
hospital_layer = gmaps.symbol_layer(hospital_geo[['lat','lng']], fill_color='black', stroke_color='black', scale=1)

inc_fig = gmaps.figure(display_toolbar=False, layout={
        'width': '700px',
        'height': '550px',
        'padding': '3px',
        'border': '1px solid black'})
inc_fig.add_layer(zcta_inc_layer)
inc_legend = widgets.HBox([widgets.HTML(lgnd.inc_legend)])
inc_fig.add_layer(hospital_layer)
inc_fig.add_layer(doctor_layer)
inc_fig = widgets.VBox([inc_fig, inc_legend])

embed_minimal_html('dmv_income.html', views=[inc_fig])

emp_fig = gmaps.figure(display_toolbar=False, layout={
        'width': '700px',
        'height': '550px',
        'padding': '3px',
        'border': '1px solid black'})
emp_fig.add_layer(zcta_emp_layer)
emp_legend = widgets.HBox([widgets.HTML(lgnd.emp_legend)])
emp_fig.add_layer(hospital_layer)
emp_fig.add_layer(doctor_layer)
emp_fig = widgets.VBox([emp_fig, emp_legend])

embed_minimal_html('dmv_employment.html', views=[emp_fig])

ins_fig = gmaps.figure(display_toolbar=False, layout={
        'width': '700px',
        'height': '550px',
        'padding': '3px',
        'border': '1px solid black'})
ins_fig.add_layer(zcta_ins_layer)
ins_legend = widgets.HBox([widgets.HTML(lgnd.ins_legend)])
ins_fig.add_layer(hospital_layer)
ins_fig.add_layer(doctor_layer)
ins_fig = widgets.VBox([ins_fig, ins_legend])

embed_minimal_html('dmv_insurance.html', views=[ins_fig])
